Refactor tests (sonar code smells)
